Extract Mask trait and impls into a different mod
