+++ src/lib.rs
@@ -13,27 +13,8 @@ impl From<std::net::AddrParseError> for NetAddrError {
 	}
 }
 
-pub trait Mask {
-	type Output;
-
-	fn mask(&self, other: &Self) -> Self;
-}
-
-impl Mask for Ipv4Addr {
-	type Output = Self;
-
-	fn mask(&self, other: &Self) -> Self::Output {
-		Self::Output::from((u32::from(*self)) & (u32::from(*other)))
-	}
-}
-
-impl Mask for Ipv6Addr {
-	type Output = Self;
-
-	fn mask(&self, other: &Self) -> Self::Output {
-		Self::Output::from((u128::from(*self)) & (u128::from(*other)))
-	}
-}
+mod mask;
+pub use mask::*;
 
 pub trait Broadcast {
 	type Output;
